requirements.txt
aiohttp
beautifulsoup4
aioresponses
pytest-asyncio
